Moved type in different files
