+++ Sources/Platform.swift
@@ -1,0 +1,17 @@
+//
+//  DeviceVersion.swift
+//
+//  Created by Inder Kumar Rathore on 01/11/20.
+//  Copyright (c) 2015. All rights reserved.
+//
+
+import Foundation
+
+public enum Platform {
+    case iPhone
+    case iPodTouch
+    case iPad
+    case appleWatch
+    case appleTV
+    case unknown
+}
